Very early version of using PHPStan for parsing.
Current run on deptrac itself finds new dependencies not found with the old method when calling `parent::method()` by resolving the `parent` to the actual class.

There is 10_000 things surely wrong with the implementation that needs fixing in the next passes, there are no tests, BUT, it works.

Big thanks to Tomas Votruba for Rector and his blog as it allowed me by example to integrate PHPStan DIC to deptrac.