To aid extension developers, I have extracted all the default Deptrac behaviour, that could be replaced/augmented by extension developers into standalone "module" called `DefaultBehavior`.
I have set as a goal that all the `DefaultBehavior` should be "implementable" by extension developers. That means is MUST NOT depend on any of the Deptrac internals. It can only depend on the exposed `Contract` and third party libraries.

This as a consequence forced me to refactor a lot of the code inside the "Core" module.

This is a good thing, because now we can confidently say that we do not expose contract that are "useless" in a sense that even though they are exposed, nobody could possibly implement extensions based on this contract, because they were missing access to some other features, that were still hidden as internal to the tool.

The result of all of this is that the Deptrac `DefaultBehavior` is just a set of extensions that we ship and autoload with the tool. Anybody could have implemented them themselves and also anybody can easily replace them.

Lastly, those `DefaultBehavior`s can serve as a reference implementations for extension developers. They can use them as a starting point for their own extensions by mimicking the code and behavior.