feat!: Expose contracts (#109)
* Add ability to define a custom mapper for the baseline store/load for people that need/want to store the baseline differently.

2 use-cases immediately come to mind:
- https://github.com/qossmic/deptrac-src/pull/45#issuecomment-2180757721
- using different format for storing violation in the PHP configuration (currently you still need to load violations from a yaml file even when using PHP config file for Deptrac)

* - Move ParserInterface.php into the Contract.
- Move AstMap related data classes into the contract.

This is in preparation to allow custom implementations of the AstParser for Deptrac purposes. This is useful for 2 reasons:
- It will allow closing https://github.com/qossmic/deptrac/issues/1402 by recommending a custom version of the parser implemented in the userspace with a supported contract to implement by the custom version
- allow for replacing the Nikic parser by PHPStan Parser.

Neither of those are implemented in this commit, but the rest of the required dependencies for the first will be exposed in the next commit.

* Added support for fully defining custom Ast Parser implementations as well as custom "Reference Extractors" for the existing Nikic PHP parser.

* To aid extension developers, I have extracted all the default Deptrac behaviour, that could be replaced/augmented by extension developers into standalone "module" called `DefaultBehavior`.

I have set as a goal that all the `DefaultBehavior` should be "implementable" by extension developers. That means is MUST NOT depend on any of the Deptrac internals. It can only depend on the exposed `Contract` and third party libraries.

This as a consequence forced me to refactor a lot of the code inside the "Core" module.

This is a good thing, because now we can confidently say that we do not expose contract that are "useless" in a sense that even though they are exposed, nobody could possibly implement extensions based on this contract, because they were missing access to some other features, that were still hidden as internal to the tool.

The result of all of this is that the Deptrac `DefaultBehavior` is just a set of extensions that we ship and autoload with the tool. Anybody could have implemented them themselves and also anybody can easily replace them.

Lastly, those `DefaultBehavior`s can serve as a reference implementations for extension developers. They can use them as a starting point for their own extensions by mimicking the code and behavior.

* Added test coverage

* Fix Nikic parser update changes not reflected in the cherry pick.
Fix after cherry pick mistakes
Formatting.
Cleanup of leftover empty files

* Documentation